Add support for X-Keywords (to/from Labels)
During IMAP fetch, the message's label names are now added to the
message "X-Keywords" header as a comma-separated list.

During IMAP append, any keywords listed in the "X-Keywords" header
will be converted to Labels (which will be created if they do not
exist) and then those labels will be assigned to the message.

Some CLI commands have also been added for testing purposes.